Flashing game screen now works correctly.

+++ headers/utilities.h
@@ -15,7 +15,7 @@ struct SDL_Renderer;
 
 	// loads an image from a given path and returns an SDL_Texture* 
  	// contains optional color arguments (for color keying that are by default set to color key to white)
-SDL_Texture* loadImage(SDL_Renderer* renderer, std::string path, bool haveColorKey = true, Uint32 R = 0xFF, Uint32 G = 0xFF, Uint32 B = 0xFF);
+SDL_Texture* loadImage(SDL_Renderer* renderer, std::string path, bool haveColorKey = true, Uint8 R = 0xFF, Uint8 G = 0xFF, Uint8 B = 0xFF);
 
 	// checks if a given point is within bounds of a given SDL_Rect
 bool isPointInRect(int x, int y, const SDL_Rect& rect);
